op3: Update how we generate our blob lists
 * Split up the lists similar to what we do for CYNGN production
   stuff so we can work with the BSP if we want to.
 * Use the new generic extract utils in vendor/cm!

+++ extract-files.sh
@@ -5,35 +5,13 @@ set -e
 export DEVICE=oneplus3
 export VENDOR=oneplus
 
-function extract() {
-    for FILE in `egrep -v '(^#|^$)' $1`; do
-        OLDIFS=$IFS IFS=":" PARSING_ARRAY=($FILE) IFS=$OLDIFS
-        FILE=`echo ${PARSING_ARRAY[0]} | sed -e "s/^-//g"`
-        DEST=${PARSING_ARRAY[1]}
-        if [ -z $DEST ]; then
-            DEST=$FILE
-        fi
-	echo "Extracting /system/$FILE ..."
-        DIR=`dirname $FILE`
-        if [ ! -d $2/$DIR ]; then
-            mkdir -p $2/$DIR
-        fi
-        if [ "$SRC" = "adb" ]; then
-            # Try CM target first
-            adb pull /system/$DEST $2/$DEST
-            # if file does not exist try OEM target
-            if [ "$?" != "0" ]; then
-                adb pull /system/$FILE $2/$DEST
-            fi
-        else
-            cp $SRC/system/$FILE $2/$DEST
-            # if file dot not exist try destination
-            if [ "$?" != "0" ]; then
-                cp $SRC/system/$DEST $2/$DEST
-            fi
-        fi
-    done
-}
+MY_DIR=$(dirname "${BASH_SOURCE[0]}")
+HELPER="$MY_DIR"/../../../vendor/cm/build/tools/extract_utils.sh
+if [ ! -f "$HELPER" ]; then
+    echo "Unable to find helper script at $HELPER"
+    exit 1
+fi
+source "$HELPER"
 
 if [ $# -eq 0 ]; then
   SRC=adb
@@ -51,9 +29,13 @@ else
   fi
 fi
 
-BASE=../../../vendor/$VENDOR/$DEVICE/proprietary
-rm -rf $BASE/*
+export SRC
 
-extract ../../$VENDOR/$DEVICE/proprietary-files.txt $BASE
+BASE="$MY_DIR"/../../../vendor/$VENDOR/$DEVICE/proprietary
+rm -rf "${BASE:?}/"*
 
-./setup-makefiles.sh
+extract "$MY_DIR"/proprietary-files-qc.txt "$BASE"
+extract "$MY_DIR"/proprietary-files-qc-perf.txt "$BASE"
+extract "$MY_DIR"/proprietary-files.txt "$BASE"
+
+"$MY_DIR"/setup-makefiles.sh
